♻️ create ContributionWithGIthubData projection and refactor logic inside
