+++ marketplace-domain/src/contribution/projections/github/mod.rs
@@ -1,0 +1,5 @@
+mod github;
+pub use github::*;
+
+#[cfg(test)]
+mod test;
