+++ marketplace-domain/src/event_projection.rs
@@ -1,0 +1,5 @@
+use crate::EventAggregate;
+
+pub trait Projection<A: EventAggregate>: Send + Sync {
+	fn project(&self, event: &A::Event);
+}
