+++ marketplace-indexer/src/domain/obervers/mod.rs
@@ -14,8 +14,8 @@ use marketplace_domain::*;
 
 #[automock]
 pub trait Observer: Send + Sync {
-	fn on_connect(&self, indexer_id: &IndexerId);
-	fn on_new_event(&self, event: &Event);
-	fn on_new_block(&self, block_hash: &BlockHash);
-	fn on_reorg(&self);
+	fn on_connect(&self, _indexer_id: &IndexerId) {}
+	fn on_new_event(&self, _event: &Event) {}
+	fn on_new_block(&self, _block_hash: &BlockHash) {}
+	fn on_reorg(&self) {}
 }
